back/requirements.txt
Django==5.0.3
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
uvicorn==0.29.0
